feature/header-css
-deleted unnecessary files
-installed react-bootstrap/added boostrap theme file from bootswatch
- changed Navbar Component to Header Component
