+++ client/package.json
@@ -10,6 +10,7 @@
     "js-cookie": "^3.0.1",
     "materialize-css": "^1.0.0-rc.2",
     "react": "^17.0.2",
+    "react-bootstrap": "^1.6.4",
     "react-dom": "^17.0.2",
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
