+++ client/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import 'materialize-css/dist/css/materialize.min.css';
+import './bootstrap.min.css'
+// import 'materialize-css/dist/css/materialize.min.css';
 
 import App from './App';
 import reportWebVitals from './reportWebVitals';
